red img

+++ fractol.h
@@ -6,7 +6,7 @@
 /*   By: kgriset <kgriset@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/27 12:03:53 by kgriset           #+#    #+#             */
-/*   Updated: 2024/03/01 15:04:57 by kgriset          ###   ########.fr       */
+/*   Updated: 2024/03/01 15:30:39 by kgriset          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,9 +34,9 @@ typedef struct s_vars {
     void *mlx;
     void *win;
     int p_x;
-    int offset_x;
+    double offset_x;
     int p_y;
-    int offset_y;
+    double offset_y;
     int zoom;
     double x;
     double y;
